requirements.txt
fastapi==0.115.0
Pillow==10.4.0
Requests==2.32.3
uvicorn==0.31.0
jinja2==3.1.4
python-multipart=0.0.12
